requirements.txt
requests==2.7.0
PyYAML==5.4.1